Refatora crawler e modifica links desatualizados
