+++ mpsp/.gitignore
@@ -1,0 +1,4 @@
+mpsp
+__pycache__
+output/
+
